requirements.txt
anthropic==0.7.4
flask==2.3.2
tavily-python==0.3.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2

